class for contour info, return type change for Controller prototype
heading1 - moved the storage of contours, moments and centers into a class created for the purpose, then rewrote the code to integrate this new class. The intent being to allow the entire mess to be sorted.
Controller - return type switched from bool to uint8_t according to the idea that tasks may wish to inform the controller what task to choose next. Not intended to compile.