Changed Task to abstract
While previously, a "task" was simply a wrapper for a function pointer and an instance name (which made it hard to associate persistent data with a task), this new structure requires each "task" to be derived from the abstract Task class and instantiated once when registered with the TaskManager instance. This should make it much easier to store scoped persistent data on a task-by-task basis.