Significant restructuring of program for Teensy.
The intent of this restructuring was to minimize programmer effort required to interface data with the interpreter. As such, where possible, wrappers were built to allow common data and functions to be packaged into a std::map. std::unordered_map would have been preferred, but the STL port being used at the moment doesn't support it. The significance of std::map is that elements can be accessed directly using strings as keys.
Global variables have been minimized, and although access speed is probably reduced, the interpreter code is much cleaner through the use of direct string access and/or loops in almost all cases.

The command structure has been extended for Config and Pilot mode, and an additional mode Test has been added. Additionally, a global disable has been added in the form of op_mode, in preparation for integration with a kill switch.

I would consider this a good stepping off point, and so I intend for this to be the last commit on this file.
Development will continue from this point with Robosub_Jul28_2018.ino