implement experimental feature: 'persistent' Threadable
A Threadable (regardless of RunLevel) has the option to call makePersistent() in its init() function at which point calls to queueCleanUp() will have no effect. In essence, the thread will never be unloaded until the program ends. This type of behavior is useful for threads that access a system resource in a blocking manner and cannot be reasonably expected to give up this resource at an arbitrary time. However, it is then up to the programmer to keep other threads from accessing this resource as behavior will be undefined. NOTE - This feature has not been tested!